Migrate tube arb to commonMain with kotlinx-datetime
Stations come from the generated tubeStationsData list (no runtime
CSV parsing). Date math switches from java.time.LocalDateTime to
kotlinx.datetime.LocalDateTime with Instant-based arithmetic in UTC.
The Station data class already lives in commonMain (Station.kt);
this commit covers FareMethod, Journey, tubeStation, and tubeJourney.
No test pins a specific Journey date, so the random output drift is
acceptable.
